requirements.txt
dill
gpustat==0.6.0
h5py==3.1.0
imutils==0.5.4
jupyter==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
matplotlib
matplotlib-inline 
numpy 
panda==0.3.1
pandas==1.3.4
Pillow==9.0.1
scipy
six
tensorflow-gpu==2.6.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-probability==0.14.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tqdm
opencv-python